+++ src/main/java/com/wskang/trip/repository/PlanRepository.java
@@ -3,5 +3,5 @@ package com.wskang.trip.repository;
 import com.wskang.trip.entity.Plan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanRepository extends JpaRepository<Plan, Integer> {
+public interface PlanRepository extends JpaRepository<Plan, Long> {
 }
