+++ src/main/java/com/wskang/trip/repository/SubPlanRepository.java
@@ -3,5 +3,5 @@ package com.wskang.trip.repository;
 import com.wskang.trip.entity.SubPlan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubPlanRepository extends JpaRepository<SubPlan, Integer> {
+public interface SubPlanRepository extends JpaRepository<SubPlan, Long> {
 }
